+++ src/components/LicenseViewer/index.js
@@ -1,0 +1,1 @@
+export { default as LicenseViewer } from './LicenseViewer';
