+++ src/components/OBSVerses/OBSContent.js
@@ -1,18 +1,14 @@
 import React, { useEffect, useState, useContext } from 'react';
 
-import { Box, Button } from '@material-ui/core';
+import { Box } from '@material-ui/core';
 import { useTranslation } from 'react-i18next';
 import { useProjector } from '@texttree/projector-mode-rcl';
 
 import { ReferenceContext, AppContext } from '../../context';
-import { ContextMenu } from '../ContextMenu';
+import { ContextMenu, LicenseViewer } from '../../components';
 import { useScrollToVerse } from '../../hooks';
 
-import axios from 'axios';
-import ReactMarkdown from 'react-markdown';
-
 import useStyles from './style';
-import { DialogUI } from '../DialogUI';
 
 const initialPositionContextMenu = {
   left: null,
@@ -32,8 +28,7 @@ function OBSContent({
   const [positionContextMenu, setPositionContextMenu] = useState(
     initialPositionContextMenu
   );
-  const [license, setLicense] = useState('');
-  const [openModal, setOpenModal] = useState(false);
+
   const {
     state: { referenceSelected },
     actions: { setReferenceBlock },
@@ -57,17 +52,6 @@ function OBSContent({
   const [verseRef] = useScrollToVerse('center');
   const { setData } = useProjector();
 
-  const getLicense = () => {
-    const { owner, name } = resource;
-    try {
-      axios
-        .get(`${server}/${owner}/${name}/raw/branch/master/LICENSE.md`)
-        .then((res) => setLicense(res.data))
-        .catch((err) => console.log(err));
-    } catch (error) {
-      console.log(error);
-    }
-  };
   const mdToVerses = (md) => {
     let _markdown = md.replaceAll('\u200B', '').split(/\n\s*\n\s*/);
     const headerMd = _markdown.shift().trim().slice(1);
@@ -169,14 +153,14 @@ function OBSContent({
           <br />
           <i>{linkMd}</i>
           <br />
-          <Button
-            onClick={() => {
-              setOpenModal(true);
-              getLicense();
+          <LicenseViewer
+            config={{
+              server,
+              repository: resource?.name,
+              owner: resource?.owner,
+              branch: 'master',
             }}
-          >
-            {t('License')}
-          </Button>
+          />
         </>
       );
 
@@ -192,14 +176,7 @@ function OBSContent({
   return (
     <>
       {verses}
-      <DialogUI
-        open={openModal}
-        maxWidth={'sm'}
-        onClose={() => setOpenModal(false)}
-        title={`License`}
-      >
-        <ReactMarkdown className={'md'}>{license}</ReactMarkdown>
-      </DialogUI>
+
       <ContextMenu position={positionContextMenu} setPosition={setPositionContextMenu} />
     </>
   );
